+++ tests/window_test.cc
@@ -142,3 +142,17 @@ TEST(Window, focused_widget) {
 
   widget_destroy(w);
 }
+
+TEST(Window, strongly_focus) {
+  value_t v1;
+  value_t v2;
+  widget_t* w = window_create(NULL, 10, 20, 30, 40);
+
+  value_set_bool(&v1, TRUE);
+  ASSERT_EQ(WINDOW_BASE(w)->strongly_focus, FALSE);
+  ASSERT_EQ(widget_set_prop(w, WIDGET_PROP_STRONGLY_FOCUS, &v1), RET_OK);
+
+  ASSERT_EQ(widget_get_prop(w, WIDGET_PROP_STRONGLY_FOCUS, &v2), RET_OK);
+  ASSERT_EQ(WINDOW_BASE(w)->strongly_focus, TRUE);
+  ASSERT_EQ(WINDOW_BASE(w)->strongly_focus, value_bool(&v2));
+}